service-examseating-python/requirements.txt
Flask
mysql-connector-python
python-dotenv
pandas